backend/requirements.txt
fastapi
uvicorn
pyyaml
faker>=20.0.0
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
pyarrow>=14.0.0
openpyxl>=3.1.0
